+++ packages/api/src/db/schema/index.ts
@@ -57,6 +57,7 @@ export * from './conductStrikes';
 export * from './contacts';
 export * from './deviceAccountContexts';
 export * from './deviceCredentials';
+export * from './deviceJoinCodes';
 export * from './devicePrincipalBackfillConflicts';
 export * from './devicePrincipals';
 export * from './deviceSessionAccounts';